requirements.txt
fastapi
uvicorn
langchain
pydantic
python-dotenv
